Switch to tauri-v2
