+++ rustfmt.toml
@@ -1,8 +1,15 @@
 max_width = 120
 reorder_imports = true
+imports_layout = "Vertical"
+imports_granularity = "Crate"
+group_imports = "StdExternalCrate"
 ignore = [
     ".git/",
+    "dist/",
     "node_modules/",
     "src/",
-    "src-tauri/target/"
+    "src-tauri/target/",
+    "*.toml",
+    "*.json",
+    "*.lock",
 ]
