+++ src-tauri/Cargo.toml
@@ -7,29 +7,33 @@ license = ""
 repository = ""
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
 
 [build-dependencies]
-tauri-build = { version = "1.5", features = [] }
+tauri-build = { version = "2.0.0-rc", features = [] }
 
 [dependencies]
-tauri = { version = "1.5", features = ["shell-open"] }
+tauri = { version = "2.0.0-rc", features = [] }
+tauri-plugin-log = { version = "2.0.0-rc" }
+tauri-plugin-shell = "2.0.0-rc"
+
+tokio = { version = "1.39", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-reqwest = { version = "0.11.23", features = ["json"] }
-tokio = { version = "1.35.1", features = ["full"] }
-diesel = { version = "^2.1.0", features = [
+reqwest = { version = "0.12", features = ["json"] }
+libsqlite3-sys = { version = "0.29", features = ["bundled"] }
+diesel = { version = "2.2", features = [
     "sqlite",
 ], git = "https://github.com/diesel-rs/diesel/" }
-diesel_migrations = { version = "^2.1.0", features = [
+diesel_migrations = { version = "2.2", features = [
     "sqlite",
 ], git = "https://github.com/diesel-rs/diesel" }
-libsqlite3-sys = { version = "0.27.0", features = ["bundled"] }
-dirs = "5.0.1"
-chrono = "0.4.34"
-typeshare = "1.0.1"
-tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
-log = "0.4.20"
+dirs = "5.0"
+chrono = "0.4"
+typeshare = "1.0"
+log = "0.4"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem
