+++ src-tauri/gen/schemas/capabilities.json
@@ -1,0 +1,14 @@
+{
+    "migrated": {
+        "identifier": "migrated",
+        "description": "permissions that were migrated from v1",
+        "local": true,
+        "windows": [
+            "main"
+        ],
+        "permissions": [
+            "core:default",
+            "shell:allow-open"
+        ]
+    }
+}
