+++ src/components/BankAccountDataProvider/Select.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { invoke } from "@tauri-apps/api/tauri";
+import { invoke } from "@tauri-apps/api/core";
 import { Select, Option } from "@material-tailwind/react";
 
 interface BankingProviderSelectProps {
