+++ src/components/BankAccounts/Select.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { invoke } from "@tauri-apps/api/tauri";
+import { invoke } from "@tauri-apps/api/core";
 import { Select, Option } from "@material-tailwind/react";
 import { BankInfo } from "../../models/typeshare_definitions";
